+++ packages/plugin/src/Freeform.php
@@ -1115,16 +1115,19 @@ class Freeform extends Plugin
 
     private function initBundles()
     {
-        $classMap = ClassMapGenerator::createMap(__DIR__.'/Bundles');
-        foreach ($classMap as $class => $path) {
-            $reflectionClass = new \ReflectionClass($class);
-            if (
-                $reflectionClass->implementsInterface(BundleInterface::class)
-                && !$reflectionClass->isAbstract()
-                && !$reflectionClass->isInterface()
-            ) {
-                $reflectionClass->newInstance();
+        static $initialized;
+
+        if (null === $initialized) {
+            $classMap = ClassMapGenerator::createMap(__DIR__.'/Bundles');
+            foreach ($classMap as $class => $path) {
+                $reflectionClass = new \ReflectionClass($class);
+                if ($reflectionClass->implementsInterface(BundleInterface::class) && !$reflectionClass->isAbstract(
+                    ) && !$reflectionClass->isInterface()) {
+                    $reflectionClass->newInstance();
+                }
             }
+
+            $initialized = true;
         }
     }
 }