+++ packages/plugin/src/Library/DataObjects/FormRenderObject/JavascriptObject.php
@@ -2,20 +2,33 @@
 
 namespace Solspace\Freeform\Library\DataObjects\FormRenderObject;
 
+use craft\helpers\Html;
 use craft\web\View;
 
 class JavascriptObject extends AbstractFormRenderObject
 {
+    private $position;
+
+    private $options;
+
+    public function __construct($value, array $replacements = [], int $position = View::POS_END, array $options = [])
+    {
+        parent::__construct($value, $replacements);
+
+        $this->position = $position;
+        $this->options = $options;
+    }
+
     /**
      * Attach the object to view.
      */
     public function attachToView()
     {
-        \Craft::$app->view->registerJs($this->getValue(), View::POS_END);
+        \Craft::$app->view->registerScript($this->getValue(), $this->position, $this->options);
     }
 
     public function getOutput(): string
     {
-        return '<script>'.$this->getValue().'</script>';
+        return Html::script($this->getValue(), $this->options);
     }
 }